+++ scripts/macros/2024/feats/unarmedFighting.js
@@ -1,0 +1,11 @@
+import {fightingStyleUnarmedFighting as unarmedFightingLegacy, fightingStyleUnarmedFightingUnarmedStrike as unarmedFightingUnarmedStrikeLegacy} from '../../../legacyMacros.js';
+export let unarmedFighting = {
+    ...unarmedFightingLegacy,
+    version: '1.2.36',
+    rules: 'modern'
+};
+export let unarmedFightingUnarmedStrike = {
+    ...unarmedFightingUnarmedStrikeLegacy,
+    version: unarmedFighting.version,
+    rules: unarmedFighting.rules
+}